Refactor keystores to use common file management utility module
